refactor route handlers into controllers
