+++ server/routes/studentRoutes.js
@@ -1,49 +1,34 @@
 var express = require('express');
 var router = express.Router();
-var Sequelize = require('Sequelize');
-var sequelize = require('../db.js').database;
-var Models = require('../db.js').Models;
-var Utils = require('../utilities.js');
-var helpers = require('../helpers.js');
+var Auth = require('../controllers/authController.js');
+var StudentController = require('../controllers/studentController.js');
 
-// MODELS
-var Class = Models.Class;
-var Teacher = Models.Teacher;
-var Student = Models.Student;
+/* Gets a student
+ * input  -> {}
+ * output -> {firstName, lastName, username, password}
+ */
 
+router.get('/:id', StudentController.getStudent);
 
-// gets a student
-router.get('/:id', helpers.checkIf('Teacher'), function (req, res) {
-  Student
-    .findById(req.params.id)
-    .then(helpers.sendResponse(res));
-});
+/* Creates a student
+ * input  -> {firstName, lastName, username, password}
+ * output -> {id, firstName, lastName, username, password, createdAt, updatedAt}
+ */
 
-// create a student
-router.post('/', function (req, res) {
-  Student
-    .create(req.body)
-    .then(helpers.sendResponse(res));
-});
+router.post('/', Auth.checkIf('Teacher'), StudentController.createStudent);
 
-// edit a student
-router.put('/:id', helpers.checkIf('Teacher'), function (req, res) {
-  Student
-    .upsert(Utils.extend(req.body, {id: req.params.id}))
-    .then(function () {
-      return Student.findById(req.params.id);
-    })
-    .then(helpers.sendResponse(res));
-});
+/* Edits a student
+ * input  -> {firstName, lastName, username, password}
+ * output -> {id, firstName, lastName, username, password, createdAt, updatedAt}
+ */
 
-// delete a student
-router.delete('/:id', helpers.checkIf('Teacher'), function (req, res) {
-  Student
-    .findById(req.params.id)
-    .then(function (student) {
-      return student.destroy();
-    })
-    .then(helpers.sendResponse(res));
-});
+router.put('/:id', Auth.checkIf('Teacher'), StudentController.editStudent);
+
+/* Deletes a student
+ * input  -> {}
+ * output -> {id, firstName, lastName, username, password, createdAt, updatedAt} 
+ */
+
+router.delete('/:id', Auth.checkIf('Teacher'), StudentController.deleteStudent);
 
 module.exports = router;
