+++ server/utilities.js
@@ -11,32 +11,5 @@ module.exports = {
       }
     }
     return object;
-  },
-
-  sendResponse: function (res) {
-    return function (data) {
-      res.status(200).send(data);
-    };
-  },
-
-  findUsername: function (username) {
-    return Promise
-      .some([
-        Models.Student.findOne({
-          where: {username: username}
-        }),
-        Models.Teacher.findOne({
-          where: {username: username}
-        })
-      ], 2)
-  },
-
-  // ROUTE HANDLERS
-  checkIf: function (model) {
-    return function (req, res, next) {
-      if (req.session.passport.user.accountType === model) { 
-        next();
-      }
-    };
   }
 };
